Add operators to SPA inputs/outputs.
This moves Network back to its own file, but extracts out the vocab
registry to the AST module to avoid circular dependencies. The
declare_input and declare_output methods dynamically add the
SpaOperatorMixin to the inheritance list.

Furthermore, this removes some methods on the AST Sink and Module nodes
that are no longer required as parsing is not done from strings anymore.
