+++ nengo_spa/__init__.py
@@ -1,7 +1,6 @@
 from nengo_spa import version
 
 from nengo_spa.actions import Actions
-from nengo_spa.ast import Network
 from nengo_spa.ast import route, ifmax
 from nengo_spa.ast import DotProduct as dot
 from nengo_spa.ast import Reinterpret as reinterpret
@@ -18,7 +17,7 @@ from nengo_spa.modules import (
     State,
     Thalamus,
     Transcode)
-from nengo_spa.network import create_inhibit_node
+from nengo_spa.network import create_inhibit_node, Network
 from nengo_spa.vocab import Vocabulary, VocabularyMap
