+++ nengo_spa/modules/compare.py
@@ -1,8 +1,8 @@
-import numpy as np
-
 import nengo
 from nengo.params import Default, IntParam
-from nengo_spa.ast import Network
+import numpy as np
+
+from nengo_spa.network import Network
 from nengo_spa.vocab import VocabularyOrDimParam
