+++ nengo_spa/modules/product.py
@@ -1,6 +1,7 @@
 import nengo
 from nengo.params import Default, IntParam
-from nengo_spa.ast import Network
+
+from nengo_spa.network import Network
 
 
 class Product(Network):
