+++ nengo_spa/modules/thalamus.py
@@ -3,10 +3,11 @@ import numpy as np
 import nengo
 from nengo.dists import Uniform
 from nengo.params import Default, IntParam, NumberParam
-from nengo_spa.ast import Network
+from nengo.synapses import Lowpass, SynapseParam
+
 from nengo_spa.modules.scalar import Scalar
 from nengo_spa.modules.state import State
-from nengo.synapses import Lowpass, SynapseParam
+from nengo_spa.network import Network
 
 
 class Thalamus(Network):
